chore: refactor jobs
