--- frontend/src/container/InfraMonitoringK8s/Jobs/JobDetails/Events/JobEvents.styles.scss
@@ -1,289 +1,0 @@
-.job-events-container {
-	margin-top: 1rem;
-
-	.filter-section {
-		flex: 1;
-
-		.ant-select-selector {
-			border-radius: 2px;
-			border: 1px solid var(--bg-slate-400) !important;
-			background-color: var(--bg-ink-300) !important;
-
-			input {
-				font-size: 12px;
-			}
-
-			.ant-tag .ant-typography {
-				font-size: 12px;
-			}
-		}
-	}
-
-	.job-events-header {
-		display: flex;
-		justify-content: space-between;
-		gap: 8px;
-
-		padding: 12px;
-		border-radius: 3px;
-		border: 1px solid var(--bg-slate-500);
-	}
-
-	.job-events {
-		margin-top: 1rem;
-
-		.virtuoso-list {
-			overflow-y: hidden !important;
-
-			&::-webkit-scrollbar {
-				width: 0.3rem;
-				height: 0.3rem;
-			}
-
-			&::-webkit-scrollbar-track {
-				background: transparent;
-			}
-
-			&::-webkit-scrollbar-thumb {
-				background: var(--bg-slate-300);
-			}
-
-			&::-webkit-scrollbar-thumb:hover {
-				background: var(--bg-slate-200);
-			}
-
-			.ant-row {
-				width: fit-content;
-			}
-		}
-
-		.skeleton-container {
-			height: 100%;
-			padding: 16px;
-		}
-	}
-
-	.ant-table {
-		.ant-table-thead > tr > th {
-			padding: 12px;
-			font-weight: 500;
-			font-size: 12px;
-			line-height: 18px;
-
-			background: rgb(18, 19, 23);
-			border-bottom: none;
-
-			color: var(--Vanilla-400, #c0c1c3);
-			font-family: Inter;
-			font-size: 11px;
-			font-style: normal;
-			font-weight: 600;
-			line-height: 18px; /* 163.636% */
-			letter-spacing: 0.44px;
-			text-transform: uppercase;
-
-			&::before {
-				background-color: transparent;
-			}
-		}
-
-		.ant-table-thead > tr > th:has(.jobname-column-header) {
-			background: var(--bg-ink-400);
-		}
-
-		.ant-table-cell {
-			padding: 12px;
-			font-size: 13px;
-			line-height: 20px;
-			color: var(--bg-vanilla-100);
-			background: rgb(18, 19, 23);
-			border-bottom: none;
-		}
-
-		.ant-table-cell:has(.jobname-column-value) {
-			background: var(--bg-ink-400);
-		}
-
-		.jobname-column-value {
-			color: var(--bg-vanilla-100);
-			font-family: 'Geist Mono';
-			font-style: normal;
-			font-weight: 600;
-			line-height: 20px; /* 142.857% */
-			letter-spacing: -0.07px;
-		}
-
-		.status-cell {
-			.active-tag {
-				color: var(--bg-forest-500);
-				padding: 4px 8px;
-				border-radius: 4px;
-				font-size: 12px;
-				font-weight: 500;
-			}
-		}
-
-		.progress-container {
-			.ant-progress-bg {
-				height: 8px !important;
-				border-radius: 4px;
-			}
-		}
-
-		.ant-table-tbody > tr:hover > td {
-			background: rgba(255, 255, 255, 0.04);
-		}
-
-		.ant-table-cell:first-child {
-			text-align: justify;
-		}
-
-		.ant-table-cell:nth-child(2) {
-			padding-left: 16px;
-			padding-right: 16px;
-		}
-
-		.ant-table-cell:nth-child(n + 3) {
-			padding-right: 24px;
-		}
-		.column-header-right {
-			text-align: right;
-		}
-		.ant-table-tbody > tr > td {
-			border-bottom: none;
-		}
-
-		.ant-table-thead
-			> tr
-			> th:not(:last-child):not(.ant-table-selection-column):not(.ant-table-row-expand-icon-cell):not([colspan])::before {
-			background-color: transparent;
-		}
-
-		.ant-empty-normal {
-			visibility: hidden;
-		}
-	}
-
-	.ant-pagination {
-		position: fixed;
-		bottom: 0;
-		width: calc(100% - 64px);
-		background: rgb(18, 19, 23);
-		padding: 16px;
-		margin: 0;
-
-		// this is to offset intercom icon till we improve the design
-		padding-right: 72px;
-
-		.ant-pagination-item {
-			border-radius: 4px;
-
-			&-active {
-				background: var(--bg-robin-500);
-				border-color: var(--bg-robin-500);
-
-				a {
-					color: var(--bg-ink-500) !important;
-				}
-			}
-		}
-	}
-}
-
-.job-events-list-container {
-	flex: 1;
-	height: calc(100vh - 272px) !important;
-	display: flex;
-	height: 100%;
-
-	.raw-log-content {
-		width: 100%;
-		text-wrap: inherit;
-		word-wrap: break-word;
-	}
-}
-
-.job-events-list-card {
-	width: 100%;
-	margin-top: 12px;
-
-	.ant-table-wrapper {
-		height: 100%;
-		overflow-y: auto;
-
-		&::-webkit-scrollbar {
-			width: 0.3rem;
-			height: 0.3rem;
-		}
-
-		&::-webkit-scrollbar-track {
-			background: transparent;
-		}
-
-		&::-webkit-scrollbar-thumb {
-			background: var(--bg-slate-300);
-		}
-
-		&::-webkit-scrollbar-thumb:hover {
-			background: var(--bg-slate-200);
-		}
-
-		.ant-row {
-			width: fit-content;
-		}
-	}
-
-	.ant-card-body {
-		padding: 0;
-
-		height: 100%;
-		width: 100%;
-	}
-}
-
-.logs-loading-skeleton {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	gap: 8px;
-	padding: 8px 0;
-
-	.ant-skeleton-input-sm {
-		height: 18px;
-	}
-}
-
-.no-logs-found {
-	height: 50vh;
-	width: 100%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-
-	padding: 24px;
-	box-sizing: border-box;
-
-	.ant-typography {
-		display: flex;
-		align-items: center;
-		gap: 16px;
-	}
-}
-
-.lightMode {
-	.filter-section {
-		border-top: 1px solid var(--bg-vanilla-300);
-		border-bottom: 1px solid var(--bg-vanilla-300);
-
-		.ant-select-selector {
-			border-color: var(--bg-vanilla-300) !important;
-			background-color: var(--bg-vanilla-100) !important;
-			color: var(--bg-ink-200);
-		}
-	}
-}
-
-.periscope-btn-icon {
-	cursor: pointer;
-}
