--- frontend/src/container/InfraMonitoringK8s/Jobs/JobDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import JobEvents from './JobEvents';
-
-export default JobEvents;
