--- frontend/src/container/InfraMonitoringK8s/Jobs/JobDetails/Logs/index.ts
@@ -1,3 +1,0 @@
-import JobLogs from './JobLogsDetailedView';
-
-export default JobLogs;
