--- frontend/src/container/InfraMonitoringK8s/Jobs/JobDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import JobMetrics from './JobMetrics';
-
-export default JobMetrics;
