--- frontend/src/container/InfraMonitoringK8s/Jobs/JobDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import JobTraces from './JobTraces';
-
-export default JobTraces;
